+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.0.5] - 30/7/2020
+
+- Ability to pass a Custom Painter.
+- Added `MagnifierPainter` and `CrosshairMagnifierPainter`.
+- Upadted example to show use of Custom Painter.
+- Added more documentation.
+
 ## [0.0.4] - 30/7/2020
 
 - Fix Typo